+++ docs/conf.py
@@ -1,5 +1,5 @@
 project = "scrapyrt"
-copyright = "%Y, Zyte Group Ltd"
+project_copyright = "%Y, Zyte Group Ltd"
 author = "Zyte Group Ltd"
 version = "0.16"
 release = "0.16.0"